documentation and some refactoring
