+++ Code/VFSPrototype/VFSBase/Implementation/FileSystemTextManipulator.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using VFSBase.Exceptions;
 using VFSBase.Interfaces;
 
 namespace VFSBase.Implementation
